atlas-ml/requirements.txt
torch
torchvision
timm
Pillow
matplotlib
scikit-learn
numpy
